# openapi-spec.yaml
openapi: 3.0.3
info:
  title: GigaID API
  description: |
    # GigaID Decentralized Identity and Credential Management System

    GigaID is a decentralized identity system that combines DIDs (Decentralized Identifiers), 
    Verifiable Credentials, Ethereum blockchain integration, and hierarchical organizational 
    structures to manage educational credentials and authentication.

    ## Key Features
    - **Hierarchical Organizations**: Giga → Country Office → Government → School
    - **Verifiable Credentials**: Issue, verify, and revoke educational credentials
    - **Multiple Authentication Methods**: WebAuthn, password-based, and hybrid authentication
    - **Blockchain Integration**: Ethereum wallet management and transactions
    - **API Key Authentication**: Third-party integration support
    - **JWT-based Sessions**: Secure token-based authentication

        ## Authentication

    The API uses JWT (JSON Web Token) based authentication:

    1. **JWT Bearer Token**: Include `Authorization: Bearer <token>` header
    2. **Session-based**: Traditional session cookies (legacy WebAuthn only)

  version: 1.0.0
  contact:
    name: UNICEF Giga Initiative
    url: https://giga.global
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://gigaidpoc-production.up.railway.app
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoints
  schemas:
...
    UserRegistration:
      type: object
      required: [username, email, displayName, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: "john_doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        displayName:
          type: string
          minLength: 1
          maxLength: 100
          example: "John Doe"
        password:
          type: string
          minLength: 6
          maxLength: 100
          example: "secure123456"
    UserLogin:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: "john_doe"
        password:
          type: string
          example: "secure123456"
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "User registered successfully"
        user:
          $ref: "#/components/schemas/User"
        auth:
          $ref: "#/components/schemas/AuthTokens"
    AuthTokens:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIs..."
        refreshToken:
          type: string
          description: JWT refresh token
          example: "eyJhbGciOiJIUzI1NiIs..."
        tokenType:
          type: string
          example: "Bearer"
        expiresIn:
          type: string
          example: "24h"
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        username:
          type: string
          example: "john_doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        displayName:
          type: string
          example: "John Doe"
        did:
          type: string
          example: "did:ethr:0x..."
        authMethod:
          type: string
          enum: [webauthn, password, hybrid]
          example: "password"
        organizationId:
          type: string
          format: uuid
          nullable: true
        organizationRole:
          type: string
          enum: [admin, staff, member]
          nullable: true

# Security applied to all endpoints by default (JWT Bearer Token)
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      description: Register a new user with password authentication and receive JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
        "409":
          description: Username already exists
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and receive JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid credentials
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Use refresh token to get new access token
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Token refreshed successfully

  # Organization Management Endpoints
...
  /api/hierarchy/hierarchy:
    get:
      tags: [Organizations]
      summary: Get complete organization hierarchy
      description: Retrieve the complete organizational hierarchy starting from Giga root
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Hierarchy retrieved successfully
...
  /api/hierarchy/credentials/issue:
    post:
      tags: [Credentials]
      summary: Issue a verifiable credential
      description: Issue a verifiable credential between organizations in the hierarchy
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [issuerDid, subjectDid, credentialType, claims]
              properties:
                issuerDid:
                  type: string
                  example: "did:ethr:0x..."
                subjectDid:
                  type: string
                  example: "did:ethr:0x..."
                credentialType:
                  type: string
                  enum:
                    [
                      country_office_authorization,
                      government_authorization,
                      school_authorization,
                      student_enrollment,
                      information_worker,
                      attendance_certificate,
                      graduation_diploma,
                    ]
                  example: "government_authorization"
                claims:
                  type: object
                  additionalProperties: true
                  example:
                    authorizedRegions: ["Kenya"]
                    validFrom: "2025-06-26T13:22:08Z"
                    authorityLevel: "National"
                expiresAt:
                  type: string
                  format: date-time
                  description: "Optional expiration date"
      responses:
        "201":
          description: Credential issued successfully
...
  /api/hierarchy/credentials/verify:
    post:
      tags: [Credentials]
      summary: Verify a credential
      description: Verify the authenticity and validity of a verifiable credential
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [credential]
              properties:
                credential:
                  type: string
                  description: "JWT-encoded verifiable credential"
      responses:
        "200":
          description: Credential verification result
  /api/hierarchy/credentials/issued/{did}:
    get:
      tags: [Credentials]
      summary: Get credentials issued by DID
      description: Retrieve all credentials issued by a specific DID
      security:
        - bearerAuth: []
      parameters:
        - name: did
          in: path
          required: true
          schema:
            type: string
            example: "did:ethr:0x..."
      responses:
        "200":
          description: Issued credentials retrieved successfully
  /api/hierarchy/credentials/received/{did}:
    get:
      tags: [Credentials]
      summary: Get credentials received by DID
      description: Retrieve all credentials received by a specific DID
      security:
        - bearerAuth: []
      parameters:
        - name: did
          in: path
          required: true
          schema:
            type: string
            example: "did:ethr:0x..."
      responses:
        "200":
          description: Received credentials retrieved successfully

  # User Management Endpoints
...
  /api/balance:
    get:
      tags: [Blockchain]
      summary: Get user's ETH balance
      description: Retrieve the current Ethereum balance for the authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Balance retrieved successfully
  /api/send-transaction:
    post:
      tags: [Blockchain]
      summary: Send Ethereum transaction
      description: Send an Ethereum transaction from the user's wallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, amount]
              properties:
                to:
                  type: string
                  minLength: 42
                  maxLength: 42
                  example: "0x742d35Cc6635C0532925a3b8D400900f"
                amount:
                  type: string
                  example: "0.1"
                gasLimit:
                  type: string
                  example: "21000"
                gasPrice:
                  type: string
                  example: "20000000000"
      responses:
        "200":
          description: Transaction sent successfully
  /api/transaction/{txHash}:
    get:
      tags: [Blockchain]
      summary: Get transaction status
      description: Retrieve the status of a specific transaction
      security:
        - bearerAuth: []
      parameters:
        - name: txHash
          in: path
          required: true
          schema:
            type: string
            example: "0x1234567890abcdef..."
      responses:
        "200":
          description: Transaction status retrieved successfully

  # System Information Endpoints
...
  /api/did-preview:
    post:
      tags: [System]
      summary: Generate DID preview
      description: Generate a preview of a DID for a given username without creating the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
                  example: "preview_user"
      responses:
        "200":
          description: DID preview generated successfully

  # Legacy WebAuthn Endpoints removed